[FEATURE] Google Analytics integration and metrics
Implements Google Analytics account authentication with OAuth 2.0 using
a new backend module. Profiles can be assigned to sites in that module.

Metrics data is available via a ``GoogleAnalytics`` data source that
can be used in generic data views inside the inspector.

The ``Document`` node type is enhanced with a new analytics tab and
some default views to display common metrics.

Stats and views can be customized in Settings.yaml and NodeTypes.yaml.

Resolves: NEOS-679
Change-Id: Ic3c0d0efb995a2526121b0a9951585858a449887