[TASK] Add button to disconnect (logout) an account, better errors in UI
If the Google account needs to be changed or does not have correct
permissions the new button allows to disconnect that account by
removing the existing tokens. This can be triggered from the index
view or from an error message.

Additionally an error message from API calls (some Google specific
errors) will be returned as an error to the widget (instead of an
internal server error).

Change-Id: I604259ad90d60c4b531b96110a7ea7226311db48
Releases: master