[TASK] Use ga:uniquePageviews instead of ga:sessions for page metrics
Google Analytics will only count a session metric for the first page
that was visited during a session. So if we filter for a specific page
this number will be much lower than the actual sessions for that page
during the period. The metric "ga:uniquePageviews" is what we are
interested in and resembles sessions for a specific page.

Change-Id: I02e936617e257fd5e118dffbab8df9304a988cc2
Releases: master